refactor(BA-6630): move session-creation value types to the data layer
Move ImageInfo, ContainerUserInfo, ScalingGroupNetworkInfo and
DeploymentContext (with its UserContext/ContainerUserContext/ImageContext/
ResolvedPresetValues helpers) from repositories/scheduler/types/session_creation
to data/session/creation, so the sokovan scheduling-controller rule modules
depend on data/ instead of repositories/ and leave the cross-layer import cycle.
