+++ src/ai/backend/manager/repositories/deployment/db_source/db_source.py
@@ -89,6 +89,13 @@ from ai.backend.manager.data.image.types import ImageIdentifier
 from ai.backend.manager.data.model_serving.types import AppProxyRouteEntry
 from ai.backend.manager.data.permission.types import RBACElementRef
 from ai.backend.manager.data.resource.types import ScalingGroupProxyTarget
+from ai.backend.manager.data.session.creation import (
+    ContainerUserContext,
+    DeploymentContext,
+    ImageContext,
+    ResolvedPresetValues,
+    UserContext,
+)
 from ai.backend.manager.data.session.types import SessionStatus
 from ai.backend.manager.data.vfolder.types import VFolderLocation
 from ai.backend.manager.errors.deployment import (
@@ -181,13 +188,6 @@ from ai.backend.manager.repositories.deployment.types import (
     RouteServiceDiscoveryInfo,
     RouteSessionInfo,
     RouteSessionKernelInfo,
-)
-from ai.backend.manager.repositories.scheduler.types.session_creation import (
-    ContainerUserContext,
-    DeploymentContext,
-    ImageContext,
-    ResolvedPresetValues,
-    UserContext,
 )
 from ai.backend.manager.repositories.scheduling_history.creators import (
     DeploymentHistoryCreatorSpec,
