+++ src/ai/backend/manager/repositories/scheduler/types/session_creation.py
@@ -2,14 +2,10 @@
 
 from collections.abc import Mapping
 from dataclasses import dataclass, field
-from decimal import Decimal
 from typing import Any
-from uuid import UUID
 
-from ai.backend.common.docker import ImageRef
 from ai.backend.common.identifier.image import ImageID
 from ai.backend.common.types import (
-    AccessKey,
     SessionId,
     SlotName,
     SlotTypes,
@@ -17,61 +13,12 @@ from ai.backend.common.types import (
 )
 from ai.backend.manager.data.dotfile.types import DotfileBundle
 from ai.backend.manager.data.resource.types import SlotTypePolicy
+from ai.backend.manager.data.session.creation import (
+    ContainerUserInfo,
+    ImageInfo,
+    ScalingGroupNetworkInfo,
+)
 from ai.backend.manager.models.scaling_group import ScalingGroupOpts
-
-
-@dataclass
-class UserContext:
-    """User information for session creation."""
-
-    uuid: UUID
-    access_key: AccessKey
-    role: str  # UserRole as string
-    sudo_session_enabled: bool
-
-
-@dataclass
-class ContainerUserContext:
-    """Container user UID/GID information."""
-
-    uid: int | None
-    main_gid: int | None
-    supplementary_gids: list[int]
-
-
-@dataclass
-class ImageContext:
-    """Resolved image information."""
-
-    ref: ImageRef  # Image reference object
-    labels: dict[str, Any]
-
-
-@dataclass
-class ResolvedPresetValues:
-    """Resolved preset values ready for session injection."""
-
-    environ: dict[str, str]
-    args: list[str]
-
-
-@dataclass
-class DeploymentContext:
-    """Context data needed to create a session from deployment info.
-
-    Consumed by
-    :class:`ai.backend.manager.sokovan.deployment.deployment_draft_builder.DeploymentSessionDraftBuilder`
-    to assemble a :class:`SessionSpecDraft` for route-executor-driven
-    inference sessions.
-    """
-
-    created_user: UserContext
-    session_owner: UserContext
-    container_user: ContainerUserContext
-    group_id: UUID
-    resource_policy: dict[str, Any]
-    image: ImageContext
-    resolved_presets: ResolvedPresetValues | None = None
 
 
 @dataclass
@@ -83,43 +30,12 @@ class SessionDependencyData:
 
 
 @dataclass
-class ScalingGroupNetworkInfo:
-    """Network configuration from scaling group."""
-
-    use_host_network: bool
-    wsproxy_addr: str | None = None
-
-
-@dataclass
-class ImageInfo:
-    """Resolved image information."""
-
-    id: UUID
-    canonical: str
-    architecture: str
-    registry: str
-    labels: dict[str, Any]
-    # Resource spec maps slot names to {"min": value, "max": value}
-    # Values can be strings (for BinarySize), numbers, or None
-    resource_spec: dict[str, dict[str, str | int | Decimal | None]]
-
-
-@dataclass
 class AllowedScalingGroup:
     """Allowed scaling group for a user."""
 
     name: str
     is_private: bool
     scheduler_opts: ScalingGroupOpts
-
-
-@dataclass
-class ContainerUserInfo:
-    """User container UID/GID information."""
-
-    uid: int | None = None
-    main_gid: int | None = None
-    supplementary_gids: list[int] = field(default_factory=list)
 
 
 @dataclass
